refactor: Refactorizar y se cambia de nombre los components
