+++ components/comment/reaction.tsx
@@ -1,10 +1,12 @@
+import React from "react";
+
 export interface ReactionProps {
     id: number,
     type: "funny" | "sad" | "surprise"
     amount: number
 }
 
-export const Reaction = ({id, type, amount}: ReactionProps) =>{
+export const Reaction: React.FC<ReactionProps> = ({id, type, amount}: ReactionProps) => {
 
     const emoji = (): string => {
         switch (type) {
@@ -14,7 +16,8 @@ export const Reaction = ({id, type, amount}: ReactionProps) =>{
                 return '🥲';
             case "surprise":
                 return '😮';
-            default: return '😂';
+            default:
+                return '😂';
         }
     }
     return (
