+++ components/exports.ts
@@ -1,0 +1,7 @@
+export * from './comment/comment'
+export * from "./layout/header"
+export * from './ui/avatar'
+export * from './ui/input'
+export * from './ui/background-radial'
+export * from './ui/background-linear'
+export * from './comment/background-comment'
