+++ components/ui/background-radial.tsx
@@ -1,4 +1,6 @@
-export function BackgroundRadial() {
+import React from "react";
+
+export const BackgroundRadial: React.FC = () => {
     return (
         <div
             className="z-20 absolute top-0 w-full h-full bg-[radial-gradient(ellipse_at_center,_var(--tw-gradient-stops))] from-black from-0% to-transparent to-70% "/>)
